CrewAI/coder/pyproject.toml
[project]
name = "coder"
version = "0.1.0"
description = "coder using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    "crewai[tools]==1.7.2",
    "litellm>=1.0.0",
    "python-dotenv>=1.0.0",
    "crewai-tools>=0.1.0"
]

[project.scripts]
coder = "coder.main:run"
run_crew = "coder.main:run"
train = "coder.main:train"
replay = "coder.main:replay"
test = "coder.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
